+++ app/lib/bk/compat/models/steps/wait.rb
@@ -1,0 +1,20 @@
+# frozen_string_literal: true
+
+module BK
+  module Compat
+    # simple waiting step
+    class WaitStep < BaseStep
+      def to_h
+        'wait'
+      end
+
+      def <<(_other)
+        raise 'Can not add to a wait step'
+      end
+
+      def >>(_other)
+        raise 'Can not add to a wait step'
+      end
+    end
+  end
+end
